requirements.txt
requests
pytest
jsonschema
allure-pytest